fix func E at realiz_1.c

+++ include/lib.h
@@ -1,4 +1,4 @@
 #pragma once
 
-float E(int x);
+double E(int x);
 char* translation(long x);
